backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-dotenv
httpx
jinja2
